liwc dictionary words
this is the same code as the rhbDict.cxx but has only LIWC words.  As with that you must use ROOT at root.cern.ch (read those comments).  The LIWC analytic variables are not usable for this analysis for reasons explained in the paper.  "Social"  and "home" are also not usable, and "work" may suffer a corresponding problem, needs study.  This file contains a snippet on how to replace means by medians.  Plots were slightly altered in Adobce Illustrator for color scheme and editing text.